Server-side validation for the submission page.
At some point I will put some Javascript in the page so that when someone chooses February, April, June, September or November for the month, it flags up a warning if the day is out of range.

Also for the submission page HTML code, some formatting changes and also when an entry has already been provided, all other options for the <select> tags will be removed.  During testing it didn't want to use the `selected` entry even if it was marked as such.