Fix some things and cleanup.
Added mysqlclient to the requirements, this should have gone in with commit f82de1adb4de39d851090d981ef18de0d5ca7b7f

Adjusted some things with HSTS so that when Debug mode is on, it doesn't try to redirect to HTTPS

Fixed a potential bug with the Birthday model. the default for the id field was being called so would return a different UUID each time makemigrations was run, now it just returns the type and not an instance of the UUID. Migration file also included.
Being the primary key, I don't think that this would have caused any issues, but wanted to fix it anyway in case there is ever a time where for whatever reason the id field is manually updated.