Added "allow notify" checkbox to submission page and code cleanup
The submission page now has a checkbox on it to toggle notification permission, which is opt-in.

Also cleaned up the code that runs the POST portion of the submission page, using the .get() function of the MultiKeyDict things with defaults to fall back to if the entries are not there (like when submitting the form with disabled entries).
Simplified the code a bit as well, rather than adding random bits of logic to define what gets passed to the template engine's context, it's now just individual lines per section.

Resolves #3